Restructuring complete, and support movie_history update
Completely restructured lambda functions, now it is a very functional module. We now also identify history movie, and move them in to movie_history table.
